+++ CHANGES.md
@@ -35,6 +35,9 @@
 
    * Added `Documentation.SBV.Examples.Puzzles.Jugs` example, solving the water-jug transfer puzzle.
 
+   * Added `Documentation.SBV.Examples.Puzzles.AOC_2021_24` example, showing how to model an EDSL in SBV,
+     solving the advent-of-code, 2021, day 24 problem.
+
    * Added complement and difference operators for regular expressions.
    
    * Added concrete type instances of Mergeable class.